refactor: Organize the repo by physics instead of mathematical abstraction
