+++ src/serac/physics/solid/elasticity.hpp
@@ -14,7 +14,7 @@
 
 #include "mfem.hpp"
 
-#include "serac/physics/base_physics.hpp"
+#include "serac/physics/utilities/base_physics.hpp"
 
 namespace serac {
