+++ src/serac/physics/solid/hyperelastic_traction_integrator.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/integrators/hyperelastic_traction_integrator.hpp"
+#include "serac/physics/solid/hyperelastic_traction_integrator.hpp"
 
 namespace serac::mfem_ext {
