+++ src/serac/physics/thermal/thermal_conduction.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/physics/thermal_conduction.hpp"
+#include "serac/physics/thermal/thermal_conduction.hpp"
 
 #include "serac/infrastructure/logger.hpp"
 #include "serac/numerics/expr_template_ops.hpp"
