+++ tests/serac_linelastic_solver.cpp
@@ -4,7 +4,7 @@
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
-#include "serac/physics/elasticity.hpp"
+#include "serac/physics/solid/elasticity.hpp"
 
 #include <fstream>
