+++ tests/serac_wrapper_tests.cpp
@@ -12,7 +12,7 @@
 #include <gtest/gtest.h>
 #include "mfem.hpp"
 
-#include "serac/integrators/wrapper_integrator.hpp"
+#include "serac/physics/utilities/wrapper_integrator.hpp"
 
 using namespace mfem;
 using namespace serac;
